Handle callback registration for processes
Will allow us to send a signal to the generate clips process
to make it send whatever partially populated clip it is in
the middle of working on to the persist process if the camera
in question has switched off. We don't want to have clips contain
footage from two huge changes in time.